// utils/enums/school-promotion.enum.ts
export enum SchoolPromotionEnum {
    // Bachelors - 3 niveaux
    BACHELOR_DIGITAL_1 = "Bachelor Digital - 1re année",
    BACHELOR_DIGITAL_2 = "Bachelor Digital - 2e année",
    BACHELOR_DIGITAL_3 = "Bachelor Digital - 3e année",
    
    MARKETING_DIGITAL_INFLUENCE_1 = "Marketing Digital & Influence - 1re année",
    MARKETING_DIGITAL_INFLUENCE_2 = "Marketing Digital & Influence - 2e année",
    MARKETING_DIGITAL_INFLUENCE_3 = "Marketing Digital & Influence - 3e année",
    
    UX_DESIGN_PRODUCT_OWNER_1 = "UX Design & Product Owner - 1re année",
    UX_DESIGN_PRODUCT_OWNER_2 = "UX Design & Product Owner - 2e année",
    UX_DESIGN_PRODUCT_OWNER_3 = "UX Design & Product Owner - 3e année",
    
    MOTION_DESIGN_1 = "Motion Design - 1re année",
    MOTION_DESIGN_2 = "Motion Design - 2e année",
    MOTION_DESIGN_3 = "Motion Design - 3e année",
    
    UI_WEBDESIGN_1 = "UI & Webdesign - 1re année",
    UI_WEBDESIGN_2 = "UI & Webdesign - 2e année",
    UI_WEBDESIGN_3 = "UI & Webdesign - 3e année",
    
    DEVELOPPEMENT_WEB_1 = "Développement Web - 1re année",
    DEVELOPPEMENT_WEB_2 = "Développement Web - 2e année",
    DEVELOPPEMENT_WEB_3 = "Développement Web - 3e année",
    
    IA_AUTOMATION_1 = "IA & Automation - 1re année",
    IA_AUTOMATION_2 = "IA & Automation - 2e année",
    IA_AUTOMATION_3 = "IA & Automation - 3e année",
    
    // Prépa Mastère
    PREPA_MASTERE = "Prépa Mastère",
    
    // Masters - 2 niveaux
    BRAND_CONTENT_1 = "Brand Content - 1re année",
    BRAND_CONTENT_2 = "Brand Content - 2e année",
    
    DATA_CUSTOMER_EXPERIENCE_1 = "Data Customer Experience - 1re année",
    DATA_CUSTOMER_EXPERIENCE_2 = "Data Customer Experience - 2e année",
    
    DIRECTION_ARTISTIQUE_DIGITALE_1 = "Direction Artistique Digitale - 1re année",
    DIRECTION_ARTISTIQUE_DIGITALE_2 = "Direction Artistique Digitale - 2e année",
    
    LEAD_UX_1 = "Lead UX - 1re année",
    LEAD_UX_2 = "Lead UX - 2e année",
    
    TECH_LEAD_1 = "Tech Lead - 1re année",
    TECH_LEAD_2 = "Tech Lead - 2e année",
}